src/requirements.txt
pydantic>=2.5.2,<3.0.0
fastapi>=0.104.1,<1.0.0
uvicorn>=0.24.0,<1.0.0
opentelemetry-api>=1.21.0,<2.0.0
opentelemetry-propagator-gcp>=1.6.0,<2.0.0
opentelemetry-sdk>=1.21.0,<2.0.0
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-gcp-trace>=1.6.0,<2.0.0
google-cloud-firestore>=2.13.1,<3.0.0
python-dotenv>=1.1.1,<2.0.0
google-auth>=2.40.3,<3.0.0
google-cloud-pubsub>=2.31.1,<3.0.0
httpx>=0.28.1,<1.0.0
redis>=7.1.0,<8.0.0
google-cloud-tasks>=2.20.0,<3.0.0
jinja2>=3.1.6,<4.0.0
